pyproject.toml
[project]
name = "image-search"
version = "0.1.0"
description = ""
authors = [
    {name = "oceanbase", email = "open_oceanbase@oceanbase.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "towhee>=1.1.3,<2.0.0",
    "streamlit>=1.39.0,<2.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "opencv-python>=4.10.0.84,<5.0.0",
    "torchvision>=0.20.0,<1.0.0",
    "transformers>=4.45.2,<5.0.0",
    "pyobvector>=0.1.6,<1.0.0",
    "fastapi-cdn-host>=0.9.1,<1.0.0",
    "fastapi>=0.115.12,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.basedpyright]
pythonVersion = "3.10"
typeCheckingMode = "basic"
reportMissingImports = false
reportMissingModuleSource = false